+++ src/main/java/com/prography/lighton/member/domain/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.prography.lighton.member.domain;
+package com.prography.lighton.member.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
